pyproject.toml
[tool.poetry]
name = "commitly"
version = "0.1.0"
description = "Git 커밋 이후 전 과정을 자동화하는 로컬 멀티 에이전트 파이프라인"
authors = ["Commitly Team"]
readme = "README.md"
packages = [{include = "commitly", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
langgraph = "^0.0.40"
langchain = "^0.1.0"
langchain-openai = "^0.0.5"
openai = "^1.12.0"
pyyaml = "^6.0"
click = "^8.1.7"
gitpython = "^3.1.41"
psycopg2-binary = "^2.9.9"
slack-sdk = "^3.26.2"
ruff = "^0.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
black = "^24.1.0"
mypy = "^1.8.0"

[tool.poetry.scripts]
commitly = "commitly.cli.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
